+++ pynocular/__init__.py
@@ -2,8 +2,7 @@
 
 __version__ = "2.0.0rc1"
 
-from pynocular.backends.context import backend
+from pynocular.backends.context import get_backend, set_backend
 from pynocular.backends.memory import MemoryDatabaseModelBackend
 from pynocular.backends.sql import SQLDatabaseModelBackend
 from pynocular.database_model import DatabaseModel
-from pynocular.uuid_str import is_valid_uuid, UUID_STR
