+++ tests/functional/test_transactions.py
@@ -6,10 +6,10 @@ from databases import Database
 from pydantic import Field
 import pytest
 
-from pynocular.backends.context import backend, get_backend
+from pynocular.backends.context import get_backend, set_backend
 from pynocular.backends.sql import SQLDatabaseModelBackend
 from pynocular.database_model import DatabaseModel, UUID_STR
-from pynocular.db_util import create_table, drop_table, gather, transaction
+from pynocular.util import create_table, drop_table, gather, transaction
 
 logger = logging.getLogger("pynocular")
 
@@ -39,7 +39,7 @@ async def postgres_backend(postgres_database: Database):
 @pytest.mark.asyncio
 async def test_gathered_creates(postgres_backend) -> None:
     """Test that we can update the db multiple times in a gather under a single transaction"""
-    with backend(postgres_backend):
+    with set_backend(postgres_backend):
         async with get_backend().db.transaction():
             await gather(
                 Org.create(id=str(uuid4()), name="orgus borgus"),
@@ -53,7 +53,7 @@ async def test_gathered_creates(postgres_backend) -> None:
 @pytest.mark.asyncio
 async def test_gathered_updates_raise_error(postgres_backend) -> None:
     """Test that an error in one update rolls back the other when gathered"""
-    with backend(postgres_backend):
+    with set_backend(postgres_backend):
         try:
             async with get_backend().transaction():
                 await gather(
@@ -71,7 +71,7 @@ async def test_gathered_updates_raise_error(postgres_backend) -> None:
 @pytest.mark.asyncio
 async def test_serial_updates(postgres_backend) -> None:
     """Test that we can update the db serially under a single transaction"""
-    with backend(postgres_backend):
+    with set_backend(postgres_backend):
         async with get_backend().transaction():
             await Org.create(id=str(uuid4()), name="orgus borgus")
             await Org.create(id=str(uuid4()), name="porgus orgus")
@@ -83,7 +83,7 @@ async def test_serial_updates(postgres_backend) -> None:
 @pytest.mark.asyncio
 async def test_serial_updates_raise_error(postgres_backend) -> None:
     """Test that an error in one update rolls back the other when run serially"""
-    with backend(postgres_backend):
+    with set_backend(postgres_backend):
         try:
             async with get_backend().transaction():
                 await Org.create(id=str(uuid4()), name="orgus borgus")
@@ -98,7 +98,7 @@ async def test_serial_updates_raise_error(postgres_backend) -> None:
 @pytest.mark.asyncio
 async def test_nested_updates(postgres_backend) -> None:
     """Test that we can perform nested update on the db under a single transaction"""
-    with backend(postgres_backend):
+    with set_backend(postgres_backend):
         async with get_backend().transaction():
             await Org.create(id=str(uuid4()), name="orgus borgus")
 
@@ -112,7 +112,7 @@ async def test_nested_updates(postgres_backend) -> None:
 @pytest.mark.asyncio
 async def test_nested_updates_raise_error(postgres_backend) -> None:
     """Test that an error in one update rolls back the other when it is nested"""
-    with backend(postgres_backend):
+    with set_backend(postgres_backend):
         try:
             async with get_backend().transaction():
                 await Org.create(id=str(uuid4()), name="orgus borgus")
@@ -130,7 +130,7 @@ async def test_nested_updates_raise_error(postgres_backend) -> None:
 @pytest.mark.asyncio
 async def test_nested_conditional_updates_raise_error(postgres_backend) -> None:
     """Test that an error in one update rolls back the other even if its a conditional transaction"""
-    with backend(postgres_backend):
+    with set_backend(postgres_backend):
         try:
             async with get_backend().transaction():
                 await Org.create(id=str(uuid4()), name="orgus borgus")
@@ -148,7 +148,7 @@ async def test_nested_conditional_updates_raise_error(postgres_backend) -> None:
 @pytest.mark.asyncio
 async def test_open_transaction_decorator(postgres_backend) -> None:
     """Test that the open_transaction decorator will execute everything in a transaction"""
-    with backend(postgres_backend):
+    with set_backend(postgres_backend):
 
         @transaction
         async def write_than_raise_error():
@@ -167,7 +167,7 @@ async def test_open_transaction_decorator(postgres_backend) -> None:
 @pytest.mark.asyncio
 async def test_open_transaction_decorator_rolls_back(postgres_backend) -> None:
     """Test that the open_transaction decorator will roll back everything in the function"""
-    with backend(postgres_backend):
+    with set_backend(postgres_backend):
 
         @transaction
         async def write_than_raise_error():
